requirements.txt
aiofiles==24.1.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
ecdsa==0.19.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.17
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
openai==1.101.0
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyPDF2==3.0.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
zstandard==0.24.0
